* Move LandingBackground and Footerfrom navbar to StartScreen component * Move FooterBackground from footer to StartScreen component
