+++ docs/src/pages/index.js
@@ -4,6 +4,8 @@ import styles from './styles.module.css';
 import Wave from '@site/src/components/Wave';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import HomepageStartScreen from '@site/src/components/Hero/StartScreen';
+import LandingBackground from '@site/src/components/Hero/LandingBackground';
+import FooterBackground from '../components/FooterBackground';
 import Playground from '@site/src/components/Playground';
 import GestureFeatures from '@site/src/components/GestureFeatures';
 import Testimonials from '@site/src/components/Testimonials';
@@ -18,6 +20,7 @@ function Home() {
     <Layout
       title={`React Native Gesture Handler`}
       description="Declarative API exposing platform native touch and gesture system to React Native.">
+      <LandingBackground />
       <div className={styles.container}>
         <HomepageStartScreen />
         <Playground />
@@ -31,6 +34,7 @@ function Home() {
         <Sponsors />
         <HireUsSection />
       </div>
+      <FooterBackground />
     </Layout>
   );
 }
