+++ docs/src/theme/Navbar/Layout/styles.module.css
@@ -16,46 +16,3 @@
   padding-top: 2rem;
   padding: 0 1.25rem 0 1.5rem;
 }
-
-.heroBackground {
-  position: absolute;
-  top: 0;
-
-  background: linear-gradient(
-    180deg,
-    var(--swm-blue-light-40) 0%,
-    var(--swm-blue-light-20) 23%,
-    #ffffff 46%
-  );
-  width: 100%;
-  height: 130vh;
-  min-height: 70vw;
-  overflow: hidden;
-  z-index: -20;
-}
-
-[data-theme='dark'] .heroBackground {
-  background: none;
-  background-color: var(--swm-navy-dark-140);
-}
-
-@media (max-width: 996px) {
-  [class*='logoWrapperLanding'] {
-    width: 50px;
-    height: 50px;
-    display: flex;
-    justify-content: center;
-  }
-}
-
-@media (max-width: 768px) {
-  .heroBackground {
-    height: 70vh;
-  }
-}
-
-@media (max-width: 420px) {
-  .navbarLanding {
-    padding: 0 1.5rem;
-  }
-}
